# specs/023-fundacao-competitiva-sazonal/contracts/competitive-foundation.openapi.yaml
openapi: 3.1.0
info:
  title: RinhaDasLendas Competitive Foundation API
  version: 1.0.0
  description: Contrato da feature 023 para Seasons, competições, rodadas, eventos, séries, partidas mínimas e Fearless.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /temporadas:
    get:
      summary: Lista Seasons
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: estado
          schema:
            $ref: '#/components/schemas/SeasonEstado'
      responses:
        '200':
          description: Página de Seasons e contexto do calendário
          headers:
            ETag:
              description: ETag observada do CalendarioCompetitivo para transições de Season
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonPage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Cria uma Season planejada
      x-capability: CanManageSeasons
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeasonRequest'
      responses:
        '201':
          description: Season criada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /temporadas/{seasonId}:
    parameters:
      - $ref: '#/components/parameters/SeasonId'
    get:
      summary: Consulta uma Season histórica por ID sem filtro implícito
      responses:
        '200':
          description: Season encontrada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Atualiza dados editáveis de uma Season planejada
      x-capability: CanManageSeasons
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSeasonRequest'
      responses:
        '200':
          description: Season atualizada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /temporadas/{seasonId}/aberturas:
    post:
      summary: Ativa a Season e encerra atomicamente a anterior
      x-capability: CanManageSeasons
      parameters:
        - $ref: '#/components/parameters/SeasonId'
        - $ref: '#/components/parameters/IfMatchCalendar'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Season ativa e calendário atualizado
          headers:
            ETag:
              description: Nova versão do calendário
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /temporadas/{seasonId}/encerramentos:
    post:
      summary: Encerra a Season ativa
      x-capability: CanManageSeasons
      parameters:
        - $ref: '#/components/parameters/SeasonId'
        - $ref: '#/components/parameters/IfMatchCalendar'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Season encerrada
          headers:
            ETag:
              description: Nova versão do calendário
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /temporadas/{seasonId}/competicoes:
    parameters:
      - $ref: '#/components/parameters/SeasonId'
    get:
      summary: Lista competições de uma única Season
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Página de competições
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionPage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Cria competição na Season
      x-capability: CanManageCompetitions
      x-resource-condition: Presidente; ou Admin antes de existir Série iniciada no escopo; SuperAdmin negado
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompetitionRequest'
      responses:
        '201':
          description: Competição criada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /temporadas/{seasonId}/regras-publicadas:
    post:
      summary: Publica regra geral da Season para Série sem Competição
      x-capability: CanManageCompetitions
      x-resource-condition: Exclusivo do Presidente; VicePresidente pendente
      parameters:
        - $ref: '#/components/parameters/SeasonId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRulesRequest'
      responses:
        '201':
          description: Versão geral publicada
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulesVersion'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    ETag:
      description: Versão fraca do recurso, por exemplo W/"7"
      schema:
        type: string
    Location:
      description: URI canônica do recurso criado
      schema:
        type: string
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 100
    IfMatch:
      in: header
      name: If-Match
      required: true
      description: ETag do recurso; em toda rota de Partida, é a ETag do agregado Série
      schema:
        type: string
    IfMatchCalendar:
      in: header
      name: If-Match-Calendar
      required: true
      description: ETag do CalendarioCompetitivo, não da Season alvo
      schema:
        type: string
    SeasonId:
      in: path
      name: seasonId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Entrada inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Capacidade ausente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso inexistente ou oculto
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Estado, versão ou chave idempotente conflitante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [message, messageCode, errors]
      properties:
        message:
          type: string
        messageCode:
          type: string
        errors:
          type: array
          items:
            type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, messageCode, message]
      properties:
        field:
          type: string
        messageCode:
          type: string
        message:
          type: string
    SeasonEstado:
      type: string
      enum: [Planejada, Ativa, Encerrada]
...
    SeriesFormat:
      type: string
      enum: [Md3, Md5]
    DraftMode:
      type: string
      enum: [Padrao, Fearless]
    CreateSeasonRequest:
      type: object
      additionalProperties: false
      required: [nome, ano, ordemNoAno, dataInicio, dataFimExclusiva]
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 120
        ano:
          type: integer
          minimum: 2009
          maximum: 9999
        ordemNoAno:
          type: integer
          minimum: 1
        dataInicio:
          type: string
          format: date
        dataFimExclusiva:
          type: string
          format: date
    UpdateSeasonRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 120
        ano:
          type: integer
          minimum: 2009
          maximum: 9999
        ordemNoAno:
          type: integer
          minimum: 1
        dataInicio:
          type: string
          format: date
        dataFimExclusiva:
          type: string
          format: date
    SeasonSummary:
      type: object
      required: [id, nome, ano, ordemNoAno, dataInicio, dataFimExclusiva, estado, versao]
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        ano:
          type: integer
        ordemNoAno:
          type: integer
        dataInicio:
          type: string
          format: date
        dataFimExclusiva:
          type: string
          format: date
        estado:
          $ref: '#/components/schemas/SeasonEstado'
        versao:
          type: integer
          format: int64
    SeasonDetail:
      allOf:
        - $ref: '#/components/schemas/SeasonSummary'
        - type: object
          required: [quantidadeCompeticoes, acoesPermitidas]
          properties:
            quantidadeCompeticoes:
              type: integer
              minimum: 0
            ativadaEm:
              type: [string, 'null']
              format: date-time
            encerradaEm:
              type: [string, 'null']
              format: date-time
            acoesPermitidas:
              type: array
              uniqueItems: true
              items:
                type: string
    SeasonPage:
      type: object
      required: [page, pageSize, items, totalItems, totalPages, calendarioConfigurado, temporadaAtual, versaoCalendario]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/SeasonSummary'
        totalItems:
          type: integer
        totalPages:
          type: integer
        calendarioConfigurado:
          type: boolean
        temporadaAtual:
          oneOf:
            - $ref: '#/components/schemas/SeasonSummary'
            - type: 'null'
        versaoCalendario:
          type: integer
          format: int64
    CreateCompetitionRequest:
      type: object
      additionalProperties: false
      required: [nome, codigo, circuitoDiario]
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 120
        codigo:
          type: string
          minLength: 1
          maxLength: 40
        circuitoDiario:
          type: boolean
...
    PublishRulesRequest:
      type: object
      additionalProperties: false
      required: [formato, modoDraft]
      properties:
        formato:
          $ref: '#/components/schemas/SeriesFormat'
        modoDraft:
          $ref: '#/components/schemas/DraftMode'
...
    CompetitionDetail:
      type: object
      required: [id, seasonId, nome, codigo, circuitoDiario, rodadas, regrasPublicadas, versao, acoesPermitidas]
      properties:
        id:
          type: string
          format: uuid
        seasonId:
          type: string
          format: uuid
        nome:
          type: string
        codigo:
          type: string
        circuitoDiario:
          type: boolean
        rodadas:
          type: array
          items:
            $ref: '#/components/schemas/Round'
        regrasPublicadas:
          type: array
          items:
            $ref: '#/components/schemas/RulesVersion'
        versao:
          type: integer
          format: int64
        acoesPermitidas:
          type: array
          uniqueItems: true
          items:
            type: string
    CompetitionPage:
      type: object
      required: [page, pageSize, items, totalItems, totalPages, calendarioConfigurado, temporadaAtual, seasonsIncluidas]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CompetitionDetail'
        totalItems:
          type: integer
        totalPages:
          type: integer
        calendarioConfigurado:
          type: boolean
        temporadaAtual:
          oneOf:
            - $ref: '#/components/schemas/SeasonSummary'
            - type: 'null'
        seasonsIncluidas:
          type: array
          items:
            $ref: '#/components/schemas/SeasonSummary'
    Round:
      type: object
      required: [id, competicaoId, nome, ordem, versao]
      properties:
        id:
          type: string
          format: uuid
        competicaoId:
          type: string
          format: uuid
        nome:
          type: string
        ordem:
          type: integer
          minimum: 1
        versao:
          type: integer
          format: int64
    RulesVersion:
      type: object
      required: [id, seasonId, numero, formato, modoDraft, publicadaEm]
      properties:
        id:
          type: string
          format: uuid
        seasonId:
          type: string
          format: uuid
        competicaoId:
          type: [string, 'null']
          format: uuid
        numero:
          type: integer
          minimum: 1
        formato:
          $ref: '#/components/schemas/SeriesFormat'
        modoDraft:
          $ref: '#/components/schemas/DraftMode'
        publicadaEm:
          type: string
          format: date-time
...